Update lights buffer before shadow render
