+++ src/renderer/internal/batches.rs
@@ -1,9 +1,11 @@
-// renderer/batches.rs (new)
+use std::ops::Range;
+
 use crate::asset::{Handle, Mesh};
 use crate::renderer::batch::InstanceData;
 use crate::renderer::RenderPass;
 use crate::scene::components::DepthState;
 
+#[derive(Debug, Clone)]
 pub(crate) struct OrderedBatch {
     pub mesh: Handle<Mesh>,
     pub pass: RenderPass,
@@ -15,24 +17,25 @@ pub(crate) struct OrderedBatch {
 
 pub(crate) struct PreparedBatches {
     pub batches: Vec<OrderedBatch>,
-    pub opaque_range: std::ops::Range<usize>,
-    pub transparent_range: std::ops::Range<usize>,
-    pub overlay_range: std::ops::Range<usize>,
+    pub opaque_range: Range<usize>,
+    pub transparent_range: Range<usize>,
+    pub overlay_range: Range<usize>,
 }
+
 impl PreparedBatches {
-    pub fn all(&self) -> &[OrderedBatch] {
+    pub(crate) fn all(&self) -> &[OrderedBatch] {
         &self.batches
     }
 
-    pub fn opaque(&self) -> &[OrderedBatch] {
+    pub(crate) fn opaque(&self) -> &[OrderedBatch] {
         &self.batches[self.opaque_range.clone()]
     }
 
-    pub fn transparent(&self) -> &[OrderedBatch] {
+    pub(crate) fn transparent(&self) -> &[OrderedBatch] {
         &self.batches[self.transparent_range.clone()]
     }
 
-    pub fn overlay(&self) -> &[OrderedBatch] {
+    pub(crate) fn overlay(&self) -> &[OrderedBatch] {
         &self.batches[self.overlay_range.clone()]
     }
 }
