+++ src/renderer/internal/buffers.rs
@@ -1,0 +1,323 @@
+use std::mem;
+use std::num::NonZeroU64;
+
+use bytemuck::Zeroable;
+use wgpu::util::DeviceExt;
+
+use crate::renderer::internal::{OrderedBatch, RenderContext, ShadowResources};
+use crate::renderer::lights::{LightsData, LightsUniform, ShadowsUniform};
+use crate::renderer::uniforms::CameraUniform;
+
+pub(crate) struct DynamicObjectsBuffer {
+    pub(crate) buffer: wgpu::Buffer,
+    pub(crate) capacity: u32,
+    pub(crate) bind_group: wgpu::BindGroup,
+    pub(crate) bind_layout: wgpu::BindGroupLayout,
+    pub(crate) scratch: Vec<crate::renderer::ObjectData>,
+}
+
+impl DynamicObjectsBuffer {
+    pub(crate) fn new(device: &wgpu::Device, capacity: u32) -> Self {
+        let bind_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+            label: Some("ObjectsBindLayout"),
+            entries: &[wgpu::BindGroupLayoutEntry {
+                binding: 0,
+                visibility: wgpu::ShaderStages::VERTEX | wgpu::ShaderStages::FRAGMENT,
+                ty: wgpu::BindingType::Buffer {
+                    ty: wgpu::BufferBindingType::Storage { read_only: true },
+                    has_dynamic_offset: false,
+                    min_binding_size: None,
+                },
+                count: None,
+            }],
+        });
+
+        let buffer_size =
+            (capacity as usize * mem::size_of::<crate::renderer::ObjectData>()) as u64;
+        let buffer = device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("ObjectsBuffer"),
+            size: buffer_size,
+            usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::COPY_DST,
+            mapped_at_creation: false,
+        });
+
+        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
+            label: Some("ObjectsBindGroup"),
+            layout: &bind_layout,
+            entries: &[wgpu::BindGroupEntry {
+                binding: 0,
+                resource: buffer.as_entire_binding(),
+            }],
+        });
+
+        Self {
+            buffer,
+            capacity,
+            bind_group,
+            bind_layout,
+            scratch: Vec::with_capacity(capacity as usize),
+        }
+    }
+
+    pub(crate) fn update(
+        &mut self,
+        context: &RenderContext,
+        batches: &[OrderedBatch],
+    ) -> Result<(), wgpu::SurfaceError> {
+        self.scratch.clear();
+        for batch in batches {
+            self.scratch.extend(batch.instances.iter().map(|inst| {
+                crate::renderer::ObjectData::from_material(inst.transform.matrix(), &inst.material)
+            }));
+        }
+
+        let required = self.scratch.len() as u32;
+        if required > self.capacity {
+            self.grow(context, required);
+        }
+
+        if !self.scratch.is_empty() {
+            context
+                .queue
+                .write_buffer(&self.buffer, 0, bytemuck::cast_slice(&self.scratch));
+        }
+
+        Ok(())
+    }
+
+    fn grow(&mut self, context: &RenderContext, required: u32) {
+        let new_capacity = required.max(self.capacity * 2);
+        log::info!(
+            "Growing objects buffer: {} -> {}",
+            self.capacity,
+            new_capacity
+        );
+
+        let buffer_size =
+            (new_capacity as usize * mem::size_of::<crate::renderer::ObjectData>()) as u64;
+        self.buffer = context.device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("ObjectsBuffer"),
+            size: buffer_size,
+            usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::COPY_DST,
+            mapped_at_creation: false,
+        });
+
+        self.bind_group = context
+            .device
+            .create_bind_group(&wgpu::BindGroupDescriptor {
+                label: Some("ObjectsBindGroup"),
+                layout: &self.bind_layout,
+                entries: &[wgpu::BindGroupEntry {
+                    binding: 0,
+                    resource: self.buffer.as_entire_binding(),
+                }],
+            });
+
+        self.capacity = new_capacity;
+    }
+}
+
+pub(crate) struct CameraBuffer {
+    pub(crate) buffer: wgpu::Buffer,
+    pub(crate) bind_group: wgpu::BindGroup,
+    pub(crate) bind_layout: wgpu::BindGroupLayout,
+}
+
+impl CameraBuffer {
+    pub(crate) fn new(device: &wgpu::Device) -> Self {
+        let camera = CameraUniform::new();
+        let buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("CameraBuffer"),
+            contents: bytemuck::bytes_of(&camera),
+            usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
+        });
+
+        let bind_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+            label: Some("CameraBindLayout"),
+            entries: &[wgpu::BindGroupLayoutEntry {
+                binding: 0,
+                visibility: wgpu::ShaderStages::VERTEX | wgpu::ShaderStages::FRAGMENT,
+                ty: wgpu::BindingType::Buffer {
+                    ty: wgpu::BufferBindingType::Uniform,
+                    has_dynamic_offset: false,
+                    min_binding_size: Some(
+                        NonZeroU64::new(mem::size_of::<CameraUniform>() as u64).unwrap(),
+                    ),
+                },
+                count: None,
+            }],
+        });
+
+        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
+            label: Some("CameraBindGroup"),
+            layout: &bind_layout,
+            entries: &[wgpu::BindGroupEntry {
+                binding: 0,
+                resource: buffer.as_entire_binding(),
+            }],
+        });
+
+        Self {
+            buffer,
+            bind_group,
+            bind_layout,
+        }
+    }
+}
+
+pub(crate) struct LightsBuffer {
+    pub(crate) buffer: wgpu::Buffer,
+    pub(crate) shadow_buffer: wgpu::Buffer,
+    pub(crate) bind_group: wgpu::BindGroup,
+    pub(crate) bind_layout: wgpu::BindGroupLayout,
+}
+
+impl LightsBuffer {
+    pub(crate) fn new(device: &wgpu::Device, shadows: &ShadowResources) -> Self {
+        let layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+            label: Some("LightsBindLayout"),
+            entries: &[
+                wgpu::BindGroupLayoutEntry {
+                    binding: 0,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Buffer {
+                        ty: wgpu::BufferBindingType::Storage { read_only: true },
+                        has_dynamic_offset: false,
+                        min_binding_size: Some(
+                            NonZeroU64::new(mem::size_of::<LightsUniform>() as u64).unwrap(),
+                        ),
+                    },
+                    count: None,
+                },
+                wgpu::BindGroupLayoutEntry {
+                    binding: 1,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Buffer {
+                        ty: wgpu::BufferBindingType::Uniform,
+                        has_dynamic_offset: false,
+                        min_binding_size: Some(
+                            NonZeroU64::new(mem::size_of::<ShadowsUniform>() as u64).unwrap(),
+                        ),
+                    },
+                    count: None,
+                },
+                wgpu::BindGroupLayoutEntry {
+                    binding: 2,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Texture {
+                        sample_type: wgpu::TextureSampleType::Depth,
+                        view_dimension: wgpu::TextureViewDimension::D2Array,
+                        multisampled: false,
+                    },
+                    count: None,
+                },
+                wgpu::BindGroupLayoutEntry {
+                    binding: 3,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Comparison),
+                    count: None,
+                },
+                wgpu::BindGroupLayoutEntry {
+                    binding: 4,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Texture {
+                        sample_type: wgpu::TextureSampleType::Depth,
+                        view_dimension: wgpu::TextureViewDimension::D2Array,
+                        multisampled: false,
+                    },
+                    count: None,
+                },
+                wgpu::BindGroupLayoutEntry {
+                    binding: 5,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Comparison),
+                    count: None,
+                },
+                wgpu::BindGroupLayoutEntry {
+                    binding: 6,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Texture {
+                        sample_type: wgpu::TextureSampleType::Depth,
+                        view_dimension: wgpu::TextureViewDimension::D2Array,
+                        multisampled: false,
+                    },
+                    count: None,
+                },
+                wgpu::BindGroupLayoutEntry {
+                    binding: 7,
+                    visibility: wgpu::ShaderStages::FRAGMENT,
+                    ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Comparison),
+                    count: None,
+                },
+            ],
+        });
+
+        let initial = LightsUniform::zeroed();
+        let buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("LightsBuffer"),
+            contents: bytemuck::bytes_of(&initial),
+            usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::COPY_DST,
+        });
+
+        let shadow_initial = ShadowsUniform::zeroed();
+        let shadow_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("ShadowUniformBuffer"),
+            contents: bytemuck::bytes_of(&shadow_initial),
+            usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
+        });
+
+        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
+            label: Some("LightsBindGroup"),
+            layout: &layout,
+            entries: &[
+                wgpu::BindGroupEntry {
+                    binding: 0,
+                    resource: buffer.as_entire_binding(),
+                },
+                wgpu::BindGroupEntry {
+                    binding: 1,
+                    resource: shadow_buffer.as_entire_binding(),
+                },
+                wgpu::BindGroupEntry {
+                    binding: 2,
+                    resource: wgpu::BindingResource::TextureView(shadows.directional_array_view()),
+                },
+                wgpu::BindGroupEntry {
+                    binding: 3,
+                    resource: wgpu::BindingResource::Sampler(shadows.sampler()),
+                },
+                wgpu::BindGroupEntry {
+                    binding: 4,
+                    resource: wgpu::BindingResource::TextureView(shadows.spot_array_view()),
+                },
+                wgpu::BindGroupEntry {
+                    binding: 5,
+                    resource: wgpu::BindingResource::Sampler(shadows.sampler()),
+                },
+                wgpu::BindGroupEntry {
+                    binding: 6,
+                    resource: wgpu::BindingResource::TextureView(shadows.point_array_view()),
+                },
+                wgpu::BindGroupEntry {
+                    binding: 7,
+                    resource: wgpu::BindingResource::Sampler(shadows.sampler()),
+                },
+            ],
+        });
+
+        Self {
+            buffer,
+            shadow_buffer,
+            bind_group,
+            bind_layout: layout,
+        }
+    }
+
+    pub(crate) fn update(&self, queue: &wgpu::Queue, lights: &LightsData) {
+        let data = LightsUniform::from_data(lights);
+        queue.write_buffer(&self.buffer, 0, bytemuck::bytes_of(&data));
+        let shadow_data = ShadowsUniform::from_data(lights);
+
+        queue.write_buffer(&self.shadow_buffer, 0, bytemuck::bytes_of(&shadow_data));
+    }
+}
