+++ src/renderer/internal/mod.rs
@@ -1,2 +1,11 @@
-pub mod vertex;
-pub use vertex::*;
+pub mod batches;
+pub mod buffers;
+pub mod context;
+pub mod pipeline;
+pub mod shadows;
+
+pub(crate) use batches::{OrderedBatch, PreparedBatches};
+pub(crate) use buffers::{CameraBuffer, DynamicObjectsBuffer, LightsBuffer};
+pub(crate) use context::RenderContext;
+pub(crate) use pipeline::{PipelineKey, RenderPipeline, TextureBindingModel};
+pub(crate) use shadows::ShadowResources;
